Move CreateTenantRequest, UpdateTenantRequest, and TenantResponseDto to Contracts namespace
