+++ account-management/WebApi/Tenants/Contracts/CreateTenantRequest.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
 
-namespace PlatformPlatform.AccountManagement.WebApi.Tenants;
+namespace PlatformPlatform.AccountManagement.WebApi.Tenants.Contracts;
 
 [UsedImplicitly]
 public sealed record CreateTenantRequest(string Name, string Subdomain, string Email, string? Phone);
