+++ account-management/WebApi/Tenants/TenantEndpoints.cs
@@ -3,6 +3,7 @@ using MediatR;
 using PlatformPlatform.AccountManagement.Application.Tenants.Commands;
 using PlatformPlatform.AccountManagement.Application.Tenants.Queries;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
+using PlatformPlatform.AccountManagement.WebApi.Tenants.Contracts;
 using PlatformPlatform.Foundation.AspNetCoreUtils.Extensions;
 
 namespace PlatformPlatform.AccountManagement.WebApi.Tenants;
